Update API documentation link and add User API to IDeliveryPenzleClient
API documentation link in IDeliveryPenzleClient.cs was updated to point to the Penzle's GitHub repository instead of previous location. This change was made because the previous documentation was deprecated.
Also, a User API client was added to IDeliveryPenzleClient, enabling user-related operations directly through the IDeliveryPenzleClient. This will simplify and unify the operation process for users.