Feature/added (#28)
* Implement exception handling in Program.cs

Added a "NotImplementedException" to the Program.cs file. This exception will effectively manage scenarios where methods have not yet been implemented, providing clearer error reporting for debugging purposes.

* Add optional Id to CreateEntryRequest model

An optional Guid Id property was added to the CreateEntryRequest model. This addition allows for more flexibility when creating new entries, as they can now be identified by an Id if desired. This is especially useful for entries that should have a pre-defined or user-specific Id.