Removed Terminal project and added password parameter to EnrollUser method (#36)
The terminal project 'Penzle.Core.Terminal' was removed from the solution, including the related csproj and Program.cs files. This is because it was no longer necessary for the set-up of the application and was causing unwanted clutter.

In addition, a password parameter was added to the 'EnrollUser' method in 'IManagementUserClient.cs' and 'User.cs' to enhance security. This allows the user to set their own password during enrollment, rather than having a system-generated password.