Update exception type in test case
The exception type being checked for in a particular test case in ConnectionShould.cs has been updated. Previously, the test was checking if a PenzleException was being thrown. However, after reviewing the functionality, it was found that it should actually throw a BadRequestException. Hence, exception type in test case was updated to ensure it matches with the actual functionality.